+++ actions/views/modals.js
@@ -28,15 +28,3 @@ export function closeModal(modalId) {
         return {data: true};
     };
 }
-
-export function incrementAnnouncementBarCount() {
-    return {
-        type: ActionTypes.TRACK_ANNOUNCEMENT_BAR,
-    };
-}
-
-export function decrementAnnouncementBarCount() {
-    return {
-        type: ActionTypes.DISMISS_ANNOUNCEMENT_BAR,
-    };
-}
