+++ components/announcement_bar/configuration_bar/configuration_bar.jsx
@@ -14,7 +14,7 @@ import {t} from 'utils/i18n';
 
 import FormattedMarkdownMessage from 'components/formatted_markdown_message';
 
-import AnnouncementBar from '../announcement_bar.jsx';
+import AnnouncementBar from '../default_announcement_bar';
 import TextDismissableBar from '../text_dismissable_bar';
 
 const RENEWAL_LINK = 'https://licensing.mattermost.com/renew';
