+++ components/announcement_bar/text_dismissable_bar.jsx
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types';
 import {trackEvent} from 'actions/diagnostics_actions.jsx';
 import Markdown from 'components/markdown';
 
-import AnnouncementBar from './announcement_bar.jsx';
+import AnnouncementBar from './default_announcement_bar';
 
 const localStoragePrefix = '__announcement__';
 
@@ -51,7 +51,7 @@ export default class TextDismissableBar extends React.PureComponent {
     }
 
     render() {
-        if (!this.state.dismissed) {
+        if (this.state.dismissed) {
             return null;
         }
         const {allowDismissal, text, ...extraProps} = this.props;
