+++ components/announcement_bar/version_bar/version_bar.test.jsx
@@ -5,7 +5,7 @@ import React from 'react';
 import {shallow} from 'enzyme';
 
 import VersionBar from 'components/announcement_bar/version_bar/version_bar.jsx';
-import AnnouncementBar from 'components/announcement_bar/announcement_bar.jsx';
+import AnnouncementBar from 'components/announcement_bar/default_announcement_bar/announcement_bar.jsx';
 
 describe('components/VersionBar', () => {
     test('should match snapshot - bar rendered after server version change', () => {
