+++ reducers/views/index.js
@@ -4,6 +4,7 @@
 import {combineReducers} from 'redux';
 
 import admin from './admin';
+import announcementBar from './announcement_bar';
 import browser from './browser';
 import channel from './channel';
 import rhs from './rhs';
@@ -23,6 +24,7 @@ import textbox from './textbox';
 
 export default combineReducers({
     admin,
+    announcementBar,
     browser,
     channel,
     rhs,
