+++ utils/utils.jsx
@@ -1763,5 +1763,5 @@ export function getSortedUsers(reactions, currentUserId, profiles, teammateNameD
 }
 
 export function getAnnouncementBarCount(state) {
-    return state.views.modals.modalState.announcementBarCount;
+    return state.views.announcementBar.announcementBarState.announcementBarCount;
 }
